requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.0+cpu
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
